creation of the navbar component
